# [Performance] Test valid path only if file info has to be retrieved