@Glandos 2022-11-07
Is there anything I can do for the failing tests? As far as I understand:
- Performance testing / performance-8.0 (pull_request) => can't find remote ref
- S3 primary storage / php8.0-objectstore_multibucket-minio (pull_request) => Failed to connect to localhost port 9000
- S3 primary storage / s3-primary-summary (pull_request) => `if true; then exit 1; fi`, so designed to fail
- continuous-integration/drone/pr => failed on acceptance-app-files-sharing but this seems to fail at random